add image

+++ WeatherApp/Models/CoordModel.swift
@@ -1,6 +1,6 @@
 import Foundation
 
 struct CoordModel : Codable {
-    let lon: Float
-    let lat: Float
+    let lon: Double
+    let lat: Double
 }
